gcp/multi-cloud-cost-optimization-billing-api-recommender/code/terraform/function_code/requirements.txt
google-cloud-billing==1.12.0
google-cloud-bigquery==3.13.0
google-cloud-pubsub==2.18.4
google-cloud-recommender==2.11.1
google-cloud-storage==2.10.0
google-cloud-monitoring==2.16.0
requests==2.31.0